+++ src/main/java/Ui.java
@@ -1,0 +1,17 @@
+public class Ui {
+
+    public static String divider() {
+        return "--------------------";
+    }
+
+    protected static void setup() {
+//        System.out.println("Working Directory = " + System.getProperty("user.dir"));
+        System.out.println("Hello from " + "DUKE");
+        System.out.println("What can I do for you?");
+        System.out.println(divider());
+    }
+
+    protected static String checkbox(Boolean mark) {
+        return mark ? "[x]" : "[ ]";
+    }
+}
